// src/api/generated/models/BookmarkStatusResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Bookmark status response
 */
export type BookmarkStatusResponse = {
    /**
     * Content ID
     */
    content_id: string;
    /**
     * Whether content is bookmarked by user
     */
    is_bookmarked: boolean;
    /**
     * Total number of user's bookmarks
     */
    bookmark_count?: number;
};